api/requirements.txt
fastapi>=0.104.0
uvicorn>=0.24.0
python-multipart>=0.0.6
pydantic>=2.5.0
pydantic-settings>=2.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.2
chromadb>=0.4.20
pypdf>=3.17.0
python-dotenv>=1.0.0
requests>=2.31.0
sqlalchemy>=2.0.0
dashscope>=1.14.0
httpx>=0.25.0
jinja2>=3.1.2
sentence-transformers>=2.2.2
duckduckgo-search>=4.1.0
langgraph>=0.0.10
